Journal #2 Full Functionality Achieved @ 29:26.08
Each new set of cubes is fed into an ArrayList, when a new set of cubes is generated, every cube in the ArrayList is destroyed.
I also remembered to save the scene before pushing things to the git this time.

"Welcome to my house! As you can see I've knocked over many chairs because I get so tilted at the towers..."
- Memphis, Tennessee