Journal #3 Checkpoint @ 35:02.18
Vectors are added to an array on click, which is used by a loop to connect each vector with drawn debug lines.
It also calculates the magnitude of the entire line with the prebuilt magnitude function.

For whatever reason, the first/zeroth index of the array is always counted as 0,0,0 even when a Debug.Log output says otherwise.